removed panels and GSAP scroll trigger to switch to a scroll-base page layout
