+++ app/home/home-page.module.scss
@@ -1,0 +1,10 @@
+/* app/home/home-page.module.scss */
+
+.homePage {
+  width: 100%;
+  background-color: var(--page-canvas);
+}
+
+.section {
+  width: 100%;
+}
