--- app/home/home-panels.module.scss
@@ -1,60 +1,0 @@
-/* app/home/home-panels.module.scss */
-
-.panelsContainer {
-  position: relative;
-  box-sizing: border-box;
-  width: 100%;
-  max-width: $layout-content-max-width;
-  margin-left: auto;
-  margin-right: auto;
-  height: 100vh;
-  /* overflow visible so stacked panels can extend; avoid clipping during slide */
-  overflow: visible;
-}
-
-.panel {
-  position: absolute;
-  inset: 0;             // top:0; right:0; bottom:0; left:0
-  width: 100%;
-  height: 100vh;
-  display: flex;        // 👈 so children can stretch
-  //align-items: stretch;
-  //justify-content: stretch;
-
-  > * {
-    flex: 1 1 auto;
-    width: 100%;
-    min-height: 100vh;  // 👈 FORCE each child to fill viewport
-  }
-}
-
-/* Last panel: ContactMe + footer share one viewport — contact scrolls, footer stays visible */
-.panelWithFooter {
-  > * {
-    min-height: 100vh;
-    max-height: 100vh;
-  }
-}
-
-.panelContactStack {
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  min-height: 0;
-  overflow: hidden;
-}
-
-.panelContactScroll {
-  flex: 1 1 auto;
-  min-height: 0;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-}
-
-.panelFooterWrap {
-  flex: 0 0 auto;
-  width: 100%;
-}
