+++ components/Footer/index.tsx
@@ -13,8 +13,6 @@ type FooterProps = {
   isDark?: boolean;
   logoFontColor?: string;
   fontColor?: string;
-  /** Less vertical padding when rendered under home ContactMe */
-  dockedInPanel?: boolean;
 };
 
 export function Footer(props: FooterProps = {}) {
@@ -53,21 +51,16 @@ export function Footer(props: FooterProps = {}) {
         window.open('https://www.linkedin.com/in/antonio-aranda-eggermont-23aa7b8/', '_blank');
     };
 
-    const docked = props.dockedInPanel === true;
-
     if (screenDevice.isMobile) {
         return (
             <div
-              style={{ paddingTop: docked ? "1rem" : "5rem" }}
+              style={{ paddingTop: "5rem" }}
               className="container"
             >
                 <div
                     className={styles['footer-container']}
                     style={{
                         borderTopColor: fontColor,
-                        ...(docked
-                          ? { marginTop: "0.35rem", marginBottom: "0.35rem" }
-                          : {}),
                     }}
                 >
                     <div className={styles['navigation-section']}>
@@ -93,18 +86,8 @@ export function Footer(props: FooterProps = {}) {
         );
     } else {
         return (
-            <div
-              className="container"
-              style={docked ? { paddingTop: "0.5rem" } : undefined}
-            >
-                <div
-                  className={styles['footer-container']}
-                  style={
-                    docked
-                      ? { marginTop: "0.35rem", marginBottom: "0.35rem" }
-                      : undefined
-                  }
-                >
+            <div className="container">
+                <div className={styles['footer-container']}>
                     <div className={styles['logo-section']}>
                         <FooterLogo color={logoFontColor} width="55px" height="20px" />
